LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/xhr/event-error-order.sub-expected.txt

ALERT: test started
ALERT: test stepped
ALERT: load ended
ALERT: finishing test

PASS XMLHttpRequest: event - error (order of events) 
